使用double 完善注释

+++ Physics2D/Settings.cs
@@ -5,8 +5,11 @@
         /// <summary>
         /// 浮点数精度
         /// </summary>
-        public static float percision = 1e-6f;
+        public static double percision = 1e-6;
 
+        /// <summary>
+        /// 最多碰撞次数
+        /// </summary>
         public static int maxContacts = 200;
     }
 }
